+++ macros/src/service/test.rs
@@ -1,5 +1,6 @@
 use aldrin::core::introspection::{Introspectable, LexicalId};
 use aldrin::core::{ServiceUuid, TypeId};
+use aldrin::service;
 use uuid::uuid;
 
 #[allow(dead_code)]
@@ -64,4 +65,46 @@ fn raw_identifiers() {
     assert_eq!(ev2.id(), 2);
     assert_eq!(ev2.name(), "ev2");
     assert_eq!(ev2.event_type(), Some(LexicalId::STRING));
+}
+
+#[test]
+fn parse_simplified_fn_item() {
+    service! {
+        #[aldrin(schema = "test", introspection)]
+        service Foo {
+            uuid = ServiceUuid(uuid!("1db9ab1c-4688-4faf-8f91-e2bd77b0ecfb"));
+            version = 1;
+
+            fn foo @ 1 = String;
+        }
+    }
+
+    let introspection = Foo::introspection();
+    assert_eq!(
+        introspection.lexical_id(),
+        LexicalId::service("test", "Foo")
+    );
+    assert_eq!(
+        introspection.type_id(),
+        TypeId(uuid!("f84b698a-a890-5c03-915c-5511adbda545"))
+    );
+    assert_eq!(Foo::lexical_id(), introspection.lexical_id());
+
+    let layout = introspection.as_service_layout().unwrap();
+    assert_eq!(layout.schema(), "test");
+    assert_eq!(layout.name(), "Foo");
+    assert_eq!(
+        layout.uuid(),
+        ServiceUuid(uuid!("1db9ab1c-4688-4faf-8f91-e2bd77b0ecfb"))
+    );
+    assert_eq!(layout.version(), 1);
+
+    let functions = layout.functions();
+    assert_eq!(functions.len(), 1);
+    let foo = functions.get(&1).unwrap();
+    assert_eq!(foo.id(), 1);
+    assert_eq!(foo.name(), "foo");
+    assert_eq!(foo.args(), None);
+    assert_eq!(foo.ok().unwrap(), LexicalId::STRING);
+    assert_eq!(foo.err(), None);
 }